# docs/swagger.yaml
definitions:
...
  usercenter.CommonResponse:
    properties:
      code:
        type: integer
      message:
        type: string
    type: object
  usercenter.DelRequest:
    properties:
      id:
        description: '@inject_tag: validate:"required"'
        type: string
      tenantID:
        type: string
    required:
    - id
    type: object
...
  usercenter.GetAllWebSiteResponse:
    properties:
      code:
        type: integer
      data:
        items:
          $ref: '#/definitions/usercenter.WebSiteInfo'
        type: array
      message:
        type: string
    type: object
...
  usercenter.GetWebSiteDetailResponse:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/usercenter.WebSiteInfo'
      message:
        type: string
    type: object
...
  usercenter.QueryWebSiteResponse:
    properties:
      code:
        type: integer
      data:
        items:
          $ref: '#/definitions/usercenter.WebSiteInfo'
        type: array
      message:
        type: string
      pages:
        type: integer
      records:
        type: integer
      total:
        type: integer
    type: object
...
  usercenter.WebSiteInfo:
    properties:
      code:
        description: 网站编号
        type: string
      id:
        type: string
      isMust:
        description: 系统必须要有的数据
        type: boolean
      logo:
        description: 网站Logo
        type: string
      name:
        description: 网站名称
        type: string
      projectID:
        description: 所属项目
        type: string
      tenantID:
        description: 租户ID
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /api/core/web_site/add:
    post:
      consumes:
      - application/json
      description: 新增
      parameters:
      - description: jwt token
        in: header
        name: authorization
        required: true
        type: string
      - description: Add WebSite
        in: body
        name: account
        required: true
        schema:
          $ref: '#/definitions/usercenter.WebSiteInfo'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/usercenter.CommonResponse'
      summary: 新增
      tags:
      - 网站配置管理
  /api/core/web_site/all:
    get:
      consumes:
      - application/json
      description: 查询所有
      parameters:
      - description: jwt token
        in: header
        name: authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/usercenter.GetAllWebSiteResponse'
      summary: 查询所有
      tags:
      - 网站配置管理
  /api/core/web_site/delete:
    delete:
      consumes:
      - application/json
      description: 删除
      parameters:
      - description: jwt token
        in: header
        name: authorization
        required: true
        type: string
      - description: Delete WebSite
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/usercenter.DelRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/usercenter.CommonResponse'
      summary: 删除
      tags:
      - 网站配置管理
  /api/core/web_site/detail:
    get:
      consumes:
      - application/json
      description: 查询明细
      parameters:
      - description: ID
        in: query
        name: id
        required: true
        type: string
      - description: jwt token
        in: header
        name: authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/usercenter.GetWebSiteDetailResponse'
      summary: 查询明细
      tags:
      - 网站配置管理
  /api/core/web_site/query:
    get:
      consumes:
      - application/json
      description: 分页查询
      parameters:
      - description: jwt token
        in: header
        name: authorization
        required: true
        type: string
      - description: 从1开始
        in: query
        name: pageIndex
        type: integer
      - description: 默认每页10条
        in: query
        name: pageSize
        type: integer
      - description: 排序字段
        in: query
        name: orderField
        type: string
      - description: 是否倒序排序
        in: query
        name: desc
        type: boolean
      - description: 网站名称
        in: query
        name: name
        type: string
      - description: 网站编号
        in: query
        name: code
        type: string
      - description: 所属项目
        in: query
        name: projectID
        type: string
      - description: 租户ID
        in: query
        name: tenantID
        type: string
      produces:
      - application/octet-stream
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/usercenter.QueryWebSiteResponse'
      summary: 分页查询
      tags:
      - 网站配置管理
  /api/core/web_site/update:
    put:
      consumes:
      - application/json
      description: 更新
      parameters:
      - description: jwt token
        in: header
        name: authorization
        required: true
        type: string
      - description: Update WebSite
        in: body
        name: account
        required: true
        schema:
          $ref: '#/definitions/usercenter.WebSiteInfo'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/usercenter.CommonResponse'
      summary: 更新
      tags:
      - 网站配置管理
...
swagger: "2.0"